+++ Robust.Shared/Prototypes/IPrototypeManager.cs
@@ -106,6 +106,10 @@ public interface IPrototypeManager
     EntityPrototype Index([ForbidLiteral] EntProtoId id);
 
     /// <inheritdoc cref="Index{T}(string)"/>
+    [Obsolete("Do not specify T when using EntProtoId")]
+    EntityPrototype Index<T>(EntProtoId protoId) where T : IInheritingPrototype;
+
+    /// <inheritdoc cref="Index{T}(string)"/>
     T Index<T>([ForbidLiteral] ProtoId<T> id) where T : class, IPrototype;
 
     /// <summary>
@@ -125,10 +129,18 @@ public interface IPrototypeManager
     bool HasIndex([ForbidLiteral] EntProtoId id);
 
     /// <inheritdoc cref="HasIndex{T}(string)"/>
+    [Obsolete("Do not specify T when using EntProtoId")]
+    bool HasIndex<T>([ForbidLiteral] EntProtoId id) where T : class, IInheritingPrototype;
+
+    /// <inheritdoc cref="HasIndex{T}(string)"/>
     bool HasIndex<T>([ForbidLiteral] ProtoId<T> id) where T : class, IPrototype;
 
     /// <inheritdoc cref="HasIndex{T}(string)"/>
     bool HasIndex([ForbidLiteral] EntProtoId? id);
+
+    /// <inheritdoc cref="HasIndex{T}(string)"/>
+    [Obsolete("Do not specify T when using EntProtoId")]
+    bool HasIndex<T>([ForbidLiteral] EntProtoId? id) where T : class, IInheritingPrototype;
 
     /// <inheritdoc cref="HasIndex{T}(string)"/>
     bool HasIndex<T>([ForbidLiteral] ProtoId<T>? id) where T : class, IPrototype;
@@ -175,6 +187,10 @@ public interface IPrototypeManager
     /// <seealso cref="TryIndex(EntProtoId,out EntityPrototype?)"/>
     bool Resolve([ForbidLiteral] EntProtoId id, [NotNullWhen(true)] out EntityPrototype? prototype);
 
+    /// <inheritdoc cref="Resolve(EntProtoId,out EntityPrototype?)"/>
+    [Obsolete("Do not specify T when using EntProtoId")]
+    bool Resolve<T>([ForbidLiteral] EntProtoId id, [NotNullWhen(true)] out EntityPrototype? prototype) where T : IInheritingPrototype;
+
     /// <summary>
     /// Resolve an <see cref="EntityPrototype"/> by ID.
     /// </summary>
@@ -199,6 +215,10 @@ public interface IPrototypeManager
     /// <returns>True if the prototype exists, false if it does not.</returns>
     /// <seealso cref="Resolve(EntProtoId,out EntityPrototype?)"/>
     bool TryIndex([ForbidLiteral] EntProtoId id, [NotNullWhen(true)] out EntityPrototype? prototype);
+
+    /// <inheritdoc cref="TryIndex(EntProtoId,out EntityPrototype?)"/>
+    [Obsolete("Do not specify T when using EntProtoId")]
+    bool TryIndex<T>([ForbidLiteral] EntProtoId id, [NotNullWhen(true)] out EntityPrototype? prototype) where T : IInheritingPrototype;
 
     /// <summary>
     /// Resolve a prototype by ID, logging an error if it does not exist.
@@ -275,6 +295,10 @@ public interface IPrototypeManager
     /// <seealso cref="TryIndex(EntProtoId?,out EntityPrototype?)"/>
     bool Resolve([ForbidLiteral] EntProtoId? id, [NotNullWhen(true)] out EntityPrototype? prototype);
 
+    /// <inheritdoc cref="Resolve(EntProtoId?,out EntityPrototype?)"/>
+    [Obsolete("Do not specify T when using EntProtoId")]
+    bool Resolve<T>([ForbidLiteral] EntProtoId? id, [NotNullWhen(true)] out EntityPrototype? prototype) where T : IInheritingPrototype;
+
     /// <summary>
     /// Resolve an <see cref="EntityPrototype"/> by ID, gracefully handling null.
     /// </summary>
@@ -299,6 +323,10 @@ public interface IPrototypeManager
     /// <returns>True if the prototype exists, false if it does not.</returns>
     /// <seealso cref="Resolve(EntProtoId?,out EntityPrototype?)"/>
     bool TryIndex([ForbidLiteral] EntProtoId? id, [NotNullWhen(true)] out EntityPrototype? prototype);
+
+    /// <inheritdoc cref="TryIndex(EntProtoId?,out EntityPrototype?)"/>
+    [Obsolete("Do not specify T when using EntProtoId")]
+    bool TryIndex<T>([ForbidLiteral] EntProtoId? id, [NotNullWhen(true)] out EntityPrototype? prototype) where T : IInheritingPrototype;
 
     /// <summary>
     /// Resolve a prototype by ID, gracefully handling null, and logging an error if it does not exist.