+++ RELEASE-NOTES.md
@@ -35,7 +35,7 @@ END TEMPLATE-->
 
 ### Breaking changes
 
-*None yet*
+* As `EntProtoId` can now be converted into both `ProtoId<EntityPrototype>` and `string`, things expecting one or the other will need an explicit cast first to avoid ambiguous invocation.
 
 ### New features